addedversion

+++ internal/app/models/companies.go
@@ -5,6 +5,6 @@ type Companies struct {
 	ID   int    `json:"id"`
 	Name string `json:"name"`
 	Slug string `json:"slug"`
-	INN  uint64 `json:"inn"`
-	KPP  uint64 `json:"kpp"`
+	INN  int    `json:"inn"`
+	KPP  int    `json:"kpp"`
 }
